engine: CodeBuilder: add section method
This allows a portion to be set aside for reuse at
some other time.

It is very useful when we expect some values to be
at certain locations but the values aren't ready, so
we just make a reference to that space and use it
when the values are ready.